pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cognitive-memory"
version = "0.1.0"
description = "Biologically-inspired cognitive memory system as an MCP server"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "mcp>=1.0.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.0",
]

[project.scripts]
cognitive-memory = "cognitive_memory.server:main"
